requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
SQLAlchemy==2.0.23

# Environment Variables
python-dotenv==1.0.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
